Remove scheduled releases.
There is no point of cutting releases from the master branch using a
schedule/cron. There is some value in terms of getting a nightly out,
but since we intend to store these packages in github - it is not going
to be easy.

We can figure out a different solution for nightlies separately.